@@ -1,115 +1,93 @@
 {
-  "name": "stromae",
-  "version": "2.1.71",
-  "description": "Web application for the management of questionnaires powered by Lunatic",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/InseeFr/Stromae.git"
-  },
-  "author": "INSEE (http://www.insee.fr)",
-  "license": "MIT",
-  "dependencies": {
-    "@craco/craco": "^6.4.0",
-    "@inseefr/lunatic": "0.3.9-experimental",
-    "@material-ui/core": "^4.11.2",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.57",
-    "@material-ui/styles": "^4.11.2",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "date-fns": "^2.17.0",
-    "keycloak-js": "16.1.1",
-    "oidc-client": "^1.10.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-error-boundary": "^3.1.3",
-    "react-markdown": "^5.0.3",
-    "react-modal": "^3.12.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.1",
-    "react-swipeable-views": "^0.13.9",
-    "react-swipeable-views-utils": "^0.14.0-alpha.0",
-    "sass": "^1.32.6",
-    "web-vitals": "^0.2.4"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "start-dev": "craco start",
-    "build": "react-scripts build && copy-and-watch configuration/* build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest",
-      "prettier",
-      "prettier/react",
-      "plugin:react-hooks/recommended",
-      "plugin:jsx-a11y/recommended",
-      "plugin:import/warnings"
-    ],
-    "plugins": [
-      "prettier",
-      "jsx-a11y"
-    ],
-    "rules": {
-      "react/prop-types": [
-        "off"
-      ],
-      "prettier/prettier": [
-        1,
-        {
-          "arrowParens": "avoid"
-        }
-      ]
-    },
-    "globals": {
-      "fetch": true,
-      "window": true,
-      "document": true,
-      "navigator": true
-    },
-    "settings": {
-      "import/resolver": {
-        "node": {
-          "paths": [
-            "src"
-          ]
-        }
-      }
-    }
-  },
-  "prettier": {
-    "singleQuote": true,
-    "trailingComma": "es5",
-    "endOfLine": "auto",
-    "arrowParens": "avoid"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "@welldone-software/why-did-you-render": "^6.2.3",
-    "copy-and-watch": "^0.1.5",
-    "eslint": "^7.20.0",
-    "eslint-config-airbnb": "18.2.1",
-    "eslint-config-prettier": "^7.1.0",
-    "eslint-plugin-import": "2.22.1",
-    "eslint-plugin-jsx-a11y": "6.4.1",
-    "eslint-plugin-prettier": "^3.3.1",
-    "eslint-plugin-react": "7.21.5",
-    "eslint-plugin-react-hooks": "4.0.0",
-    "prettier": "^2.2.1"
-  }
+	"name": "stromae",
+	"version": "2.2.0-lunaticV2",
+	"description": "Web application for the management of questionnaires powered by Lunatic",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/InseeFr/Stromae.git"
+	},
+	"author": "INSEE (http://www.insee.fr)",
+	"license": "MIT",
+	"dependencies": {
+		"@craco/craco": "^6.4.0",
+		"@inseefr/lunatic": "0.4.23-v2",
+		"@material-ui/core": "^4.11.2",
+		"@material-ui/icons": "^4.11.2",
+		"@material-ui/lab": "^4.0.0-alpha.57",
+		"@material-ui/styles": "^4.11.2",
+		"@testing-library/jest-dom": "^5.11.4",
+		"@testing-library/react": "^11.1.0",
+		"@testing-library/user-event": "^12.1.10",
+		"date-fns": "^2.17.0",
+		"keycloak-js": "16.1.1",
+		"oidc-client": "^1.10.1",
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0",
+		"react-error-boundary": "^3.1.3",
+		"react-markdown": "^5.0.3",
+		"react-modal": "^3.12.1",
+		"react-router-dom": "^5.2.0",
+		"react-scripts": "4.0.1",
+		"react-swipeable-views": "^0.13.9",
+		"react-swipeable-views-utils": "^0.14.0-alpha.0",
+		"sass": "^1.32.6",
+		"web-vitals": "^0.2.4"
+	},
+	"scripts": {
+		"start": "react-scripts start",
+		"start-dev": "craco start",
+		"build": "react-scripts build && copy-and-watch configuration/* build",
+		"test": "react-scripts test",
+		"eject": "react-scripts eject",
+		"prepare": "husky install",
+		"lint:check": "eslint . --ext .js,.jsx",
+		"lint": "npm run lint:check -- --fix",
+		"_format": "prettier '**/*.{js,jsx,json,md}'",
+		"format": "npm run _format -- --write",
+		"format:check": "npm run _format -- --list-different"
+	},
+	"lint-staged": {
+		"*.{js,jsx}": [
+			"eslint --fix"
+		],
+		"*.{js,jsx,json,md}": [
+			"prettier --write"
+		]
+	},
+	"eslintConfig": {
+		"extends": [
+			"react-app",
+			"react-app/jest"
+		]
+	},
+	"browserslist": {
+		"production": [
+			">0.2%",
+			"not dead",
+			"not op_mini all"
+		],
+		"development": [
+			"last 1 chrome version",
+			"last 1 firefox version",
+			"last 1 safari version"
+		]
+	},
+	"devDependencies": {
+		"@commitlint/cli": "^17.1.2",
+		"@commitlint/config-conventional": "^17.1.0",
+		"@welldone-software/why-did-you-render": "^6.2.3",
+		"copy-and-watch": "^0.1.5",
+		"eslint": "^8.12.0",
+		"eslint-config-airbnb": "^19.0.4",
+		"eslint-config-prettier": "^8.5.0",
+		"eslint-config-react-app": "^7.0.1",
+		"eslint-plugin-import": "^2.26.0",
+		"eslint-plugin-jsx-a11y": "^6.5.1",
+		"eslint-plugin-prettier": "^4.0.0",
+		"eslint-plugin-react": "^7.29.4",
+		"eslint-plugin-react-hooks": "^4.4.0",
+		"husky": "^8.0.1",
+		"lint-staged": "^13.0.3",
+		"prettier": "^2.7.1"
+	}
 }
