@@ -9,7 +9,7 @@ import Divider from '@material-ui/core/Divider';
 import Typography from '@material-ui/core/Typography';
 import ExpandMore from '@material-ui/icons/ExpandMore';
 import { makeStyles } from '@material-ui/core/styles';
-import { interpret } from '@inseefr/lunatic';
+import { interpret } from '@inseefr/trevas';
 import { welcomePageDictionary } from 'i18n';
 import { MarkdownTypo } from 'components/designSystem';
 import { OrchestratorContext } from 'components/orchestrator/collector';
@@ -34,11 +34,11 @@ const WelcomePage = () => {
     genericPages?.welcome || welcomePageDictionary(inseeContext);
 
   const getBodyWithVariables = (myBody, bindingDependencies) =>
-    interpret(['VTL'])({
+    interpret(myBody, {
       ...buildDefaultBindings(bindingDependencies),
       ...buildBuidings(variables),
       ...buildBuidings(personalization),
-    })(myBody);
+    });
 
   const getFinalLabel = label =>
     label || `Not yet Implemented for ${inseeContext}`;
