@@ -1,11 +1,10 @@
-import * as lunatic from '@inseefr/lunatic';
 export * from './dataDownload';
 
-export const getNotNullCollectedState = questionnaire =>
-  lunatic.getCollectedStateByValueType(questionnaire)('COLLECTED', false);
+// TODO
+export const getNotNullCollectedState = () => {};
 
 export const secureCopy = objectToCopy =>
-  JSON.parse(JSON.stringify(objectToCopy));
+  objectToCopy ? JSON.parse(JSON.stringify(objectToCopy)) : {};
 
 const getPageWithoutAnyIteration = currentPage =>
   currentPage
